fix: user can check --help even without valid URI
To create a client to make operations, we need
a valid service URI.
Before an invalid service URI exits the program
and show a message for the user to fix the URI.

This way, the user can not see the help command
information from the CLI itself even before deci-
ding to use.

I am creating the client at function level.

Creating a client for every function call is
not the right approach. But for a purpose demo,
seems reasonable.

This way the user can check the --help function
before deciding to use the tool.